Add gulpfile/browserify config to use require() in project.
This change introduces a gulpfile that watches src/*.js, and builds them to dist/ using browserify.
It will also touch an XML file of the same name in the root so that the SceneVR server triggers a
refresh. The goal is to let people more easily use require() in their SceneVR code.

As an example, I've moved the code of wave.xml to wave.js, and made use of the color package within
it.